General:  * Daily Summary Report -- If mail send fail, try to send another mail with NO attachment (but only if Eception looks like a 'message file too big'
AseTune:
 * For alarm AlarmEventClientErrorMsg, if error 1105 -- adding a table with all databases: DBName, DbSizeInMb, LogSizeInMb, LogSizeFreeInMb, LogSizeUsedInMb, LogSizeUsedPct, TransactionLogFull, DataSizeInMb, DataSizeFreeInMb, DataSizeUsedInMb, DataSizeUsedPct, DBOptions

PostgresTune:
 * CmActiveStatements -- Added columns: 'im_blocking_others_max_time_in_sec', with AlarmEventBlockingLockAlarm if above 60 seconds
 * CmActiveStatements -- Added columns: 'pid_lock_info' to show a HTML table with all locks held by this PID
 * CmActiveStatements -- Added columns: 'blocked_pids_info' to show details (SQL and LOCKS) about each PID that the PID is BLOCKING
 * CmActiveStatements -- renamed column 'query' to 'last_known_sql_statement'
 * CmActiveStatements -- Tooltip on 'wait_event_type' and 'wait_event'
 * CmPgActivity -- Tooltip on 'wait_event_type' and 'wait_event'

SqlServerTune:
 * Added alarm - AlarmEventCertificateExpiry
 * CmMemoryClerks -- Changed calculation of 'SizeMb' added 'virtual_memory_committed_kb' and 'awe_allocated_kb'
 * CmSummary -- Added 3 graphs: TargetAndTotalMemMb, MemUtilizationPct, OsMemoryFreeMb
 * CmSummary -- Added som memory columns: total_os_memory_mb, available_physical_memory_mb, system_high_memory_signal_state, system_low_memory_signal_state, Total_Server_Memory_MB, Target_Server_Memory_MB, TargetVsTotal_diff_MB, memory_used_by_sqlserver_MB, locked_pages_used_by_sqlserver_MB, process_memory_utilization_pct, process_physical_memory_low, process_virtual_memory_low
 * DBMS Config -- added SqlServerCertificates and SqlServerAgentJobs

SQL Window
 * in the tool 'ddlgen' added options '-o|--outputFile' and '-O|--overwrite'

+++ src/com/asetune/alarm/events/AlarmEventBlockingLockAlarm.java
@@ -71,7 +71,7 @@ extends AlarmEvent
 	}
 
 	/** used by CmActiveStatements */
-	public AlarmEventBlockingLockAlarm(CountersModel cm, Number threshold, int spid, Number blockingOthersMaxTimeInSec, String blockingOtherSpids, Number blockCount)
+	public AlarmEventBlockingLockAlarm(CountersModel cm, Number threshold, long spid, Number blockingOthersMaxTimeInSec, String blockingOtherSpids, Number blockCount)
 	{
 		super(
 				Version.getAppName(), // serviceType
